% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyrun_Ecodata_to_MSSPM.R
\name{convertEcodataToMSSPM}
\alias{convertEcodataToMSSPM}
\title{Keyrun Ecodata to MSSPM Environmental Covariate Conversion}
\usage{
convertEcodataToMSSPM(colNames, inputDataFrames, region, vars, outputFile)
}
\arguments{
\item{colNames}{The names of the covariate parameters that are used in MSSPM}

\item{region}{The geographical region of interest (i.e., "GB","GOM","MAB")}

\item{vars}{The description used in the data frame to specify the desired years' data}

\item{inputDataFrame}{The input Ecodata data frames}

\item{outputfile}{The output csv data file containing the MSSPM Covariate
formatted data. N.B. The csv file must begin with "Covariate_".}
}
\value{
n/a
}
\description{
Conversion of Keyrun (i.e., Ecodata) Environmental Covariate data to an MSSPM Covariate csv file
}
\examples{
remotes::install_github("NOAA-EDAB/ecodata")

colNames <- list("chloro_a","bottom_temp","sea_surface_temp")
inputDataFrames <- list(ecodata::chl_pp,ecodata::bottom_temp,ecodata::seasonal_oisst_anom)
region <- "GB" # GB, GOM, or MAB (i.e., George's Bank, Gulf of Maine, Mid-Atlantic Bight)
vars <- list("ANNUAL_PPD_MEDIAN","bottom temp anomaly in situ","Spring OISST anomaly")
outputFile <- '/home/rklasky/test/Covariate_Keyrun.csv'

convertEcodataToMSSPM(colNames,inputDataFrames,region,vars,outputFile)

}
