% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyrun_CatchCommercialAgeComp_to_MSSPM.R
\name{convertKeyrunCatchCommercialAgeCompToMSSPM}
\alias{convertKeyrunCatchCommercialAgeCompToMSSPM}
\title{Keyrun Commercial AgeComp Catch to MSSPM Catch Conversion}
\usage{
convertKeyrunCatchCommercialAgeCompToMSSPM(
  inputDataFrame,
  startYear,
  outputFile
)
}
\arguments{
\item{inputDataFrame}{The Keyrun input fishery catch data frame}

\item{startYear}{The year to start the conversion (earlier years will be skipped)}

\item{outputfile}{The output csv data file containing the MSSPM catch
formatted data. N.B. The csv file must begin with "HarvestCatch_".}
}
\value{
n/a
}
\description{
Conversion of Keyrun Commercial AgeComp Catch data file to MSSPM Catch csv file
}
\examples{
remotes::install_github("NOAA-EDAB/ms-keyrun")

inputDataFrame <- mskeyrun::realFisheryAgecomp
startYear <- 55
outputFile <- '/home/rklasky/test/HarvestCatch_Keyrun_AgeComp.csv'

convertKeyrunCatchCommercialAgeCompToMSSPM(inputDataFrame,startYear,outputFile)

}
