% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyrun_CatchCommercial_to_MSSPM.R
\name{convertKeyrunCatchCommercialToMSSPM}
\alias{convertKeyrunCatchCommercialToMSSPM}
\title{Keyrun to MSSPM Biomass Conversion}
\usage{
convertKeyrunCatchCommercialToMSSPM(
  speciesMappingDataFrame,
  catchIndex,
  outputFile
)
}
\arguments{
\item{speciesMappingDataFrame}{The data frame that contains a mapping of the NESPP3 numeric values to corresponding species names (i.e., modelName)}

\item{catchIndexDataFrame}{The input commercial catch index (i.e. commercial landings in metric tons) data frame}

\item{outputfile}{The output csv data file containing the MSSPM catch data
formatted data.  N.B. The csv file must begin with "HarvestCatch_".}
}
\value{
n/a
}
\description{
Conversion of Keyrun Commercial Catch Index file to MSSPM Catch csv file
}
\examples{
remotes::install_github("NOAA-EDAB/ms-keyrun")

speciesMappingDataFrame <- mskeyrun::focalSpecies # needed to map NESPP3 value to modelName
speciesMappingDataFrame$NESPP3 <- as.numeric(speciesMappingDataFrame$NESPP3) # necessary to eliminate the leading 0's in the NESPP3 terms

#load(file='/home/rklasky/Downloads/catchIndex.rda') # temporary until the data are released into mskeyrun
catchIndex <- mskeyrun::catchIndex

outputFile <- '/home/rklasky/test/CatchCommercial.csv'

convertKeyrunCatchCommercialToMSSPM(speciesMappingDataFrame,catchIndex,outputFile)

}
