% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyrun_CatchSim_to_MSSPM.R
\name{convertKeyrunCatchSimToMSSPM}
\alias{convertKeyrunCatchSimToMSSPM}
\title{Keyrun Simulated Catch to MSSPM Catch Conversion}
\usage{
convertKeyrunCatchSimToMSSPM(inputDataFrame, startYear, outputFile)
}
\arguments{
\item{inputDataFrame}{The Keyrun input simulated catch data frame}

\item{startYear}{The year to start the conversion (earlier years will be skipped)}

\item{outputfile}{The output csv data file containing the MSSPM catch
formatted data. N.B. The csv file must begin with "HarvestCatch_".}
}
\value{
n/a
}
\description{
Conversion of Keyrun (i.e., Atlantis) Simulated Catch Index file to MSSPM Catch csv file
}
\examples{
remotes::install_github("NOAA-EDAB/ms-keyrun")

inputDataFrame <- mskeyrun::simCatchIndex
startYear <- 55
outputFile <- '/home/rklasky/test/HarvestCatch_Keyrun_Simulated.csv'

convertKeyrunCatchSimToMSSPM(inputDataFrame,startYear,outputFile)

}
