% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyrun_ObsBM_to_MSSPM.R
\name{convertKeyrunObsBMtoMSSPM}
\alias{convertKeyrunObsBMtoMSSPM}
\title{Keyrun to MSSPM Biomass Conversion}
\usage{
convertKeyrunObsBMtoMSSPM(
  speciesMappingDataFrame,
  observedBMDataFrame,
  surveySeason,
  outputFile
)
}
\arguments{
\item{speciesMappingDataFrame}{The data frame that contains a mapping of the SVSPP numeric values to corresponding species names (i.e., modelName)}

\item{observedBMDataFrame}{The input survey index (i.e. biomass) data frame}

\item{surveySeason}{The survey season data to convert (valid options are currently FALL or SPRING)}

\item{outputfile}{The output csv data file containing the MSSPM observed biomass
formatted data.  N.B. The csv file must begin with "Biomass".}
}
\value{
n/a
}
\description{
Conversion of Keyrun Survey Index file to MSSPM Observed Biomass csv file
}
\examples{
remotes::install_github("NOAA-EDAB/ms-keyrun")

speciesMappingDataFrame <- mskeyrun::focalSpecies # needed to map SVSPP value to modelName
observedBMDataFrame <- mskeyrun::surveyIndexA4
surveySeason <- 'SPRING' # Use 'FALL' or "SPRING"
outputFile <- paste('/home/rklasky/test/BiomassAbsolute_',surveySeason,'.csv',sep="")

convertKeyrunObsBMtoMSSPM(speciesMappingDataFrame,observedBMDataFrame,surveySeason,outputFile)

}
